requirements.txt
# Core dependencies
scapy>=2.4.5
pyshark>=0.6
requests>=2.25.1
matplotlib>=3.3.0

# AI Integration
google-generativeai>=0.3.0

# Optional: For enhanced features
# pandas>=1.3.0  # For data analysis
# numpy>=1.21.0  # For numerical operations
# plotly>=5.0.0  # For interactive visualizations
# flask>=2.0.0   # For web API
# sqlite3        # Built-in Python module
